requirements.txt
streamlit
google-cloud-speech
google-cloud-texttospeech
moviepy
openai
python-dotenv